Run remove_redundant_casts refactoring

+++ src/expat_config_h.rs
@@ -1,2 +1,2 @@
 use libc::c_int;
-pub const XML_CONTEXT_BYTES: c_int = 1024 as c_int;
+pub const XML_CONTEXT_BYTES: c_int = 1024i32;
