+++ src/internal.rs
@@ -1,2 +1,2 @@
 use libc::c_int;
-pub const __INT_MAX__: c_int = 2147483647 as c_int;
+pub const __INT_MAX__: c_int = 2147483647i32;
