+++ src/stdbool_h.rs
@@ -1,3 +1,3 @@
 use libc::c_int;
-pub const true_0: c_int = 1 as c_int;
-pub const false_0: c_int = 0 as c_int;
+pub const true_0: c_int = 1i32;
+pub const false_0: c_int = 0i32;
